ai/requirements.txt
transformers
torch
sentencepiece
pdfplumber
pandas
tqdm
python-docx